refactor resolution step to separate import and symbol resolution as well as diagnostics in files contained in a resolve directory
