+++ analyzer/src/steps.rs
@@ -1,2 +1,3 @@
 pub mod collect;
 pub mod resolve;
+mod shared_diagnostics;
